requirements.txt
# local package
-e .

# external requirements
flake8
python-dotenv>=0.5.1
flet
repath
pydantic
abstractcp
polars
typing-extensions
httpx
oauthlib
rx
qrcode
cryptography
